requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt>=5.4.0
cloudinary
django-cloudinary-storage
gunicorn==25.3.0
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.12
python-decouple==3.8
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
